+++ aspecd/utils.py
@@ -104,9 +104,13 @@ class ToDictMixin:
         dictionary are those attributes accessed by getters and setters and
         hence not automatically included in the list otherwise.
 
+    .. versionchanged: 0.6.3
+        Added superclass call to preserve mro in dependent subclasses
+
     """
 
     def __init__(self):
+        super().__init__()
         if '__odict__' not in self.__dict__:
             self.__odict__ = collections.OrderedDict()
         self._exclude_from_to_dict = []