+++ tests/test_analysis.py
@@ -261,7 +261,7 @@ class TestAggregatedAnalysisStep(unittest.TestCase):
                       self.analysis.description)
 
     def test_to_dict_does_not_contain_certain_keys(self):
-        for key in ['datasets, result']:
+        for key in ['datasets', 'result']:
             with self.subTest(key=key):
                 self.assertNotIn(key, self.analysis.to_dict())