+++ docs/changelog.rst
@@ -24,6 +24,7 @@ Fixes
 * :class:`aspecd.plotting.MultiPlotter` sets axis labels when units are empty
 * :class:`aspecd.processing.Normalisation` removes unit from last axis
 * :class:`aspecd.processing.BaselineCorrection` handles zero values in range properly
+* :class:`aspecd.analysis.AggregateAnalysisStep` no longer adds ``datasets`` and ``result`` to output of ``to_dict()``
 
 
 Version 0.6.2